requirements.txt
langchain
openai
unstructured
pinecone-client
tiktoken
pdfminer.six
scikit-learn
lark